moaaa_api_services/requirements.txt
# Django
Django>=5.0,<6.0

# Django REST Framework
djangorestframework>=3.14,<4.0

# AWS SDK
boto3>=1.34,<2.0

# Environment variables
python-dotenv>=1.0,<2.0

# Database (PostgreSQL)
psycopg2-binary>=2.9,<3.0

# CORS headers
django-cors-headers>=4.3,<5.0

# Testing
pytest>=8.0,<9.0
pytest-django>=4.8,<5.0
